+++ scripts/retrieve_databundle_light.py
@@ -33,7 +33,7 @@ Each data bundle entry has the following structure:
     urls:  # list of urls by source, e.g. zenodo or google
       zenodo: {zenodo url}  # key to download data from zenodo
       gdrive: {google url}  # key to download data from google drive
-      protectedplanet: {url}  # key to download data from protected planet
+      protectedplanet: {url}  # key to download data from protected planet; the url can contain {month:s} and {year:d} to let the workflow specify the current month and year
       direct: {url}  # key to download data directly from a url; if unzip option is enabled data are unzipped
       post:  # key to download data using an url post request; if unzip option is enabled data are unzipped
         url: {url}
@@ -80,6 +80,7 @@ according to the following rules:
 - ``cutouts``: input data unzipped into the cutouts folder
 
 """
+import datetime as dt
 import os
 import re
 from zipfile import ZipFile
@@ -143,8 +144,9 @@ def download_and_unzip_zenodo(config, rootpath, hot_run=True, disable_progress=F
     """
     resource = config["category"]
     file_path = os.path.join(rootpath, "tempfile.zip")
-
+    destination = os.path.relpath(config["destination"])
     url = config["urls"]["zenodo"]
+
     if hot_run:
         try:
             logger.info(f"Downloading resource '{resource}' from cloud '{url}'")
@@ -152,7 +154,7 @@ def download_and_unzip_zenodo(config, rootpath, hot_run=True, disable_progress=F
             logger.info(f"Extracting resources")
             with ZipFile(file_path, "r") as zipObj:
                 # Extract all the contents of zip file in current directory
-                zipObj.extractall(path=config["destination"])
+                zipObj.extractall(path=destination)
             os.remove(file_path)
             logger.info(f"Downloaded resource '{resource}' from cloud '{url}'.")
         except:
@@ -187,7 +189,7 @@ def download_and_unzip_gdrive(config, rootpath, hot_run=True, disable_progress=F
     """
     resource = config["category"]
     file_path = os.path.join(rootpath, "tempfile.zip")
-
+    destination = os.path.relpath(config["destination"])
     url = config["urls"]["gdrive"]
 
     # retrieve file_id from path
@@ -225,7 +227,7 @@ def download_and_unzip_gdrive(config, rootpath, hot_run=True, disable_progress=F
         )
         with ZipFile(file_path, "r") as zipObj:
             # Extract all the contents of zip file in current directory
-            zipObj.extractall(path=config["destination"])
+            zipObj.extractall(path=destination)
 
         logger.info(f"Download resource '{resource}' from cloud '{url}'.")
 
@@ -236,7 +238,7 @@ def download_and_unzip_gdrive(config, rootpath, hot_run=True, disable_progress=F
 
 
 def download_and_unzip_protectedplanet(
-    config, rootpath, hot_run=True, disable_progress=False
+    config, rootpath, attempts=3, hot_run=True, disable_progress=False
 ):
     """
     download_and_unzip_protectedplanet(config, rootpath, dest_path,
@@ -250,6 +252,9 @@ def download_and_unzip_protectedplanet(
         Configuration data for the category to download
     rootpath : str
         Absolute path of the repository
+    attempts : int (default 3)
+        Number of attempts to download the data by month.
+        The download is attempted for the current and previous months according to the number of attempts
     hot_run : Bool (default True)
         When true the data are downloaded
         When false, the workflow is run without downloading and unzipping
@@ -262,46 +267,96 @@ def download_and_unzip_protectedplanet(
     """
     resource = config["category"]
     file_path = os.path.join(rootpath, "tempfile_wpda.zip")
-
+    destination = os.path.relpath(config["destination"])
     url = config["urls"]["protectedplanet"]
+
+    def get_first_day_of_month(date):
+        return date.replace(day=1)
+
+    def get_first_day_of_previous_month(date):
+        return get_first_day_of_month(date - dt.timedelta(days=1))
+
+    current_first_day = get_first_day_of_month(dt.datetime.today())
 
     if hot_run:
         if os.path.exists(file_path):
             os.remove(file_path)
 
-        try:
-            logger.info(f"Downloading resource '{resource}' from cloud '{url}'.")
-            progress_retrieve(url, file_path, disable_progress=disable_progress)
+        downloaded = False
 
-            zip_obj = ZipFile(file_path, "r")
+        for i in range(attempts + 1):
+            # customize url to current month
+            month_MMM = current_first_day.strftime("%b")
+            year_YYYY = current_first_day.year
+            url_iter = url.format(month=month_MMM, year=year_YYYY)
 
-            # list of zip files, which contains the shape files
-            zip_files = [
-                fname for fname in zip_obj.namelist() if fname.endswith(".zip")
-            ]
+            resource_iter = resource + " - " + month_MMM + " " + str(year_YYYY)
 
-            # extract the nested zip files
-            for fzip in zip_files:
-                # final path of the file
-                inner_zipname = os.path.join(config["destination"], fzip)
+            try:
+                logger.info(
+                    f"Downloading resource '{resource_iter}' from cloud '{url_iter}'."
+                )
+                progress_retrieve(
+                    url_iter, file_path, disable_progress=disable_progress
+                )
 
-                zip_obj.extract(fzip, path=config["destination"])
+                zip_obj = ZipFile(file_path, "r")
 
-                with ZipFile(inner_zipname, "r") as nested_zip:
-                    nested_zip.extractall(path=config["destination"])
+                # list of zip files, which contains the shape files
+                zip_files = [
+                    fname for fname in zip_obj.namelist() if fname.endswith(".zip")
+                ]
 
-                # remove inner zip file
-                os.remove(inner_zipname)
+                # if empty, the download failed
+                if not zip_files:
+                    raise Exception(
+                        "Corrupted zip file downloaded from protectedplanet"
+                    )
 
-            # remove outer zip file
-            os.remove(file_path)
+                # extract the nested zip files
+                for fzip in zip_files:
+                    # final path of the file
+                    try:
+                        inner_zipname = os.path.join(destination, fzip)
 
-            logger.info(f"Downloaded resource '{resource}' from cloud '{url}'.")
-        except:
-            logger.warning(f"Failed download resource '{resource}' from cloud '{url}'.")
-            return False
+                        zip_obj.extract(fzip, path=destination)
 
-    return True
+                        dest_nested = os.path.join(destination, fzip.split(".")[0])
+
+                        with ZipFile(inner_zipname, "r") as nested_zip:
+                            nested_zip.extractall(path=dest_nested)
+
+                        # remove inner zip file
+                        os.remove(inner_zipname)
+
+                        logger.info(f"{resource} - Successfully unzipped file '{fzip}'")
+                    except:
+                        logger.warning(
+                            f"Exception while unzipping file '{fzip}' for {resource_iter}: skipped file"
+                        )
+
+                # close and remove outer zip file
+                zip_obj.close()
+                os.remove(file_path)
+
+                logger.info(
+                    f"Downloaded resource '{resource_iter}' from cloud '{url_iter}'."
+                )
+
+                downloaded = True
+                break
+            except:
+                logger.warning(
+                    f"Failed download resource '{resource_iter}' from cloud '{url_iter}'."
+                )
+                current_first_day = get_first_day_of_previous_month(current_first_day)
+
+    if not downloaded:
+        logger.warning(
+            f"All attempts ({attempts}) to download resource '{resource}' from protected planet failed."
+        )
+
+    return downloaded
 
 
 def download_and_unpack(
@@ -381,10 +436,10 @@ def download_and_unzip_direct(config, rootpath, hot_run=True, disable_progress=F
     True when download is successful, False otherwise
     """
     resource = config["category"]
-    destination = config["destination"]
+    destination = os.path.relpath(config["destination"])
     url = config["urls"]["direct"]
 
-    file_path = os.path.join(config["destination"], os.path.basename(url))
+    file_path = os.path.join(destination, os.path.basename(url))
 
     unzip = config.get("unzip", False)
 
@@ -435,7 +490,7 @@ def download_and_unzip_hydrobasins(
     True when download is successful, False otherwise
     """
     resource = config["category"]
-    destination = config["destination"]
+    destination = os.path.relpath(config["destination"])
     url_templ = config["urls"]["hydrobasins"]["base_url"]
     suffix_list = config["urls"]["hydrobasins"]["suffixes"]
 
@@ -446,7 +501,7 @@ def download_and_unzip_hydrobasins(
 
     for rg in suffix_list:
         url = url_templ + "hybas_" + rg + "_lev" + level_code + "_v1c.zip"
-        file_path = os.path.join(config["destination"], os.path.basename(url))
+        file_path = os.path.join(destination, os.path.basename(url))
 
         all_downloaded &= download_and_unpack(
             url=url,
@@ -486,13 +541,14 @@ def download_and_unzip_post(config, rootpath, hot_run=True, disable_progress=Fal
     True when download is successful, False otherwise
     """
     resource = config["category"]
+    destination = os.path.relpath(config["destination"])
 
     # load data for post method
     postdata = config["urls"]["post"]
     # remove url feature
     url = postdata.pop("url")
 
-    file_path = os.path.join(config["destination"], os.path.basename(url))
+    file_path = os.path.join(destination, os.path.basename(url))
 
     if hot_run:
         if os.path.exists(file_path):
@@ -512,7 +568,7 @@ def download_and_unzip_post(config, rootpath, hot_run=True, disable_progress=Fal
         # then unzip it and remove the original file
         if config.get("unzip", False):
             with ZipFile(file_path, "r") as zipfile:
-                zipfile.extractall(config["destination"])
+                zipfile.extractall(destination)
 
             os.remove(file_path)
         logger.info(f"Downloaded resource '{resource}' from cloud '{url}'.")